+++ NEWS.md
@@ -1,4 +1,4 @@
-# pairwiseComparisons 0.1.2.9000
+# pairwiseComparisons 0.1.3
 
 MINOR CHANGES